requirements.txt
torch
torchaudio
lightning
pytorch_lightning
wandb
hydra-core
omegaconf
einops
numpy
jiwer